+++ ProcessChatAI.info.php
@@ -15,7 +15,7 @@
 
 $info = [
     'title' => 'AI Chatbot for ProcessWire',
-    'version' => "0.0.3Alpha",
+    'version' => "0.0.4Alpha",
     'summary' => 'Manages ChatAI chats in admin',
     'autoload' => 'admin',
     'singular' => true,